feat: 设置SessionCookieNamePrefix
